+++ services/cloudapi/struct_purchased_api_group_attribute.go
@@ -24,7 +24,7 @@ type PurchasedApiGroupAttribute struct {
 	ExpireTime     string `json:"ExpireTime" xml:"ExpireTime"`
 	RegionId       string `json:"RegionId" xml:"RegionId"`
 	BillingType    string `json:"BillingType" xml:"BillingType"`
-	InvokeTimesMax int    `json:"InvokeTimesMax" xml:"InvokeTimesMax"`
-	InvokeTimesNow int    `json:"InvokeTimesNow" xml:"InvokeTimesNow"`
+	InvokeTimesMax int64  `json:"InvokeTimesMax" xml:"InvokeTimesMax"`
+	InvokeTimesNow int64  `json:"InvokeTimesNow" xml:"InvokeTimesNow"`
 	Status         string `json:"Status" xml:"Status"`
 }
